WIP: This is a huge back and forth around probing weak-dtypes
Test suit goes down in flames, but it probably is mostly right,
many things just should not work. The biggest problem is around
comparison like operations (e.g. `abs(x - y) < value`) which
give a warning, but just need to be manually opt-in to one of
the new behaviours.

+++ numpy/core/_methods.py
@@ -179,7 +179,7 @@ def _mean(a, axis=None, dtype=None, out=None, keepdims=False, *, where=True):
     ret = umr_sum(arr, axis, dtype, out, keepdims, where=where)
     if isinstance(ret, mu.ndarray):
         ret = um.true_divide(
-                ret, rcount, out=ret, casting='unsafe', subok=False)
+                ret, rcount.astype(ret.dtype), out=ret, casting='unsafe', subok=False)
         if is_float16_result and out is None:
             ret = arr.dtype.type(ret)
     elif hasattr(ret, 'dtype'):
@@ -219,7 +219,7 @@ def _var(a, axis=None, dtype=None, out=None, ddof=0, keepdims=False, *,
         # matching rcount to arrmean when where is specified as array
         div = rcount.reshape(arrmean.shape)
     if isinstance(arrmean, mu.ndarray):
-        arrmean = um.true_divide(arrmean, div, out=arrmean, casting='unsafe',
+        arrmean = um.true_divide(arrmean, div.astype(arrmean.dtype), out=arrmean, casting='unsafe',
                                  subok=False)
     else:
         arrmean = arrmean.dtype.type(arrmean / rcount)
@@ -249,7 +249,7 @@ def _var(a, axis=None, dtype=None, out=None, ddof=0, keepdims=False, *,
     # divide by degrees of freedom
     if isinstance(ret, mu.ndarray):
         ret = um.true_divide(
-                ret, rcount, out=ret, casting='unsafe', subok=False)
+                ret, rcount.astype(ret.dtype), out=ret, casting='unsafe', subok=False)
     elif hasattr(ret, 'dtype'):
         ret = ret.dtype.type(ret / rcount)
     else:
